+++ packages/nest-auth/vite.config.ts
@@ -1,8 +1,11 @@
 /// <reference types="vitest" />
 import { defineConfig } from 'vite';
 import { resolve } from 'path';
+import { peerDependencies } from './package.json';
 import dts from 'vite-plugin-dts';
 import tsconfigPaths from 'vite-tsconfig-paths';
+
+import swc from 'unplugin-swc';
 
 export default defineConfig({
   plugins: [
@@ -13,19 +16,44 @@ export default defineConfig({
       rollupTypes: true,
     }),
     tsconfigPaths(),
+    // esbuild doesn't support a couple of features that nestjs requires, so instead
+    // we use swc. For example, see: https://github.com/nestjs/nest/issues/9228
+    swc.vite({
+      module: { type: 'es6' },
+      jsc: {
+        target: 'esnext',
+        parser: {
+          syntax: 'typescript',
+          decorators: true,
+        },
+        transform: {
+          legacyDecorator: true,
+          decoratorMetadata: true,
+        },
+        keepClassNames: true,
+        preserveAllComments: true,
+      },
+    }),
   ],
+  ssr: {
+    target: 'node',
+  },
   build: {
+    ssr: true,
+    minify: false,
     lib: {
       entry: {
         main: resolve(__dirname, 'src/main.ts'),
       },
-      name: '@spuxx/js-utils',
+      name: '@spuxx/nest-testing',
       formats: ['es'],
+    },
+    rollupOptions: {
+      external: [...Object.keys(peerDependencies)],
     },
   },
   test: {
     environment: 'node',
-    globals: true,
     silent: true,
     setupFiles: './tests/vitest.setup.ts',
     reporters: ['default', 'junit'],
@@ -33,8 +61,7 @@ export default defineConfig({
     coverage: {
       provider: 'v8',
       all: true,
-      include: ['src/**/*.ts'],
-      exclude: ['src/main.ts', 'src/**/index.ts', '**/types/**'],
+      include: ['**/*.ts'],
       reportsDirectory: 'reports/vitest/coverage',
       reporter: ['text', 'json'],
     },